review

+++ crates/cli-tools/src/action.rs
@@ -810,13 +810,11 @@ impl Display for OptLevel {
 ///
 /// This also computes the side-table and inserts it as the first section.
 pub async fn optimize_wasm(applet: impl AsRef<Path>, opt_level: Option<OptLevel>) -> Result<()> {
-    match optimize_wasm_(applet.as_ref(), opt_level).await {
-        Ok(()) => Ok(()),
-        Err(error) => {
-            let _ = fs::remove_file(applet).await;
-            Err(error)
-        }
+    let result = optimize_wasm_(applet.as_ref(), opt_level).await;
+    if result.is_err() {
+        let _ = fs::remove_file(applet).await;
     }
+    result
 }
 
 async fn optimize_wasm_(applet: impl AsRef<Path>, opt_level: Option<OptLevel>) -> Result<()> {
